pyproject.toml
[project]
name = "nexus"
version = "0.1.0"
description = "NEXUS — Nowcasting Economic eXploration Unified System"
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.0",
    "numpy>=1.24",
    "requests>=2.28",
    "httpx>=0.24",
    "pyyaml>=6.0",
    "pyarrow>=12.0",
    "statsmodels>=0.14",
    "scikit-learn>=1.3",
    "pytrends>=4.9",
    "beautifulsoup4>=4.12",
    "matplotlib>=3.7",
    "seaborn>=0.12",
    "jupyter>=1.0",
    "python-dotenv>=1.0",
    "tqdm>=4.65",
    "pytest>=7.4",
    "enahodata>=0.0.1",
]

[project.optional-dependencies]
satellite = ["earthengine-api>=0.1.370", "rasterio>=1.3"]
political = [
    "transformers>=4.40",
    "torch>=2.0",
    "anthropic>=0.25",
    "openpyxl>=3.1",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
